requirements.txt
asgiref
Django
django-filter==24.2
djangorestframework==3.15.1
sqlparse==0.5.0
typing_extensions==4.12.0
drf-yasg
